Внес все правки: экспортировал миниатюры, вызвал в main
