+++ js/data.js
@@ -1,5 +1,7 @@
 import {getId, getRandomInteger, getRandomArrayElement} from './utils.js';
 
+const PHOTO_COUNT = 25;
+
 const Like = {
   MIN: 15,
   MAX: 200
@@ -77,6 +79,6 @@ const createPhoto = () => {
   });
 };
 
-const getPhotos = (count) => Array.from({length: count}, createPhoto);
+const getPhotos = () => Array.from({length: PHOTO_COUNT}, createPhoto);
 
 export {getPhotos};
