+++ js/main.js
@@ -1,1 +1,6 @@
-import './thumbnailsCreator.js';
+import {getPhotos} from './data.js';
+import {createThumbnails} from './thumbnails.js';
+
+const photos = getPhotos();
+
+createThumbnails(photos);
